requirements.txt
flask==2.3.3
mysql-connector-python==8.2.0
gunicorn==21.2.0
